+++ CHANGELOG.rst
@@ -1,8 +1,8 @@
 Change Log
 ==========
 
-Unreleased
-----------
+2014.05.19.2
+------------
 * Improved: @joonathan added ``kwargs`` passing to the ``Blueprint`` Router
 * Improved: ``init_app`` and added ``__init__``
 * Added: Deprecation warning to ``Basic`` router